Day 17, catching up
Solution to part 2 is a little odd and a little dirty.
I definitely didn't entirely grasp the mathematical structure of
the problem, though I came close.
I think my solution might be unreasonably slow on some inputs,
but I'm not gonna check. If it is, the correct way to solve it more
efficiently would be to reset the inital_register_a to the the previous
power of 8 and search from there for the next value that matches the output.

As it stands, my solution searches intelligently until it's got most of the
input correct and then brute-forces every possible value until _all_
outputs are correct, without accounting for possible overflows into
the next output.

In my input, this only caused the search of ~30 extra entries over a perfect
solution, which is nothing at all,

But I suppose there could be cases where this covers thousands or, I suppose
on sufficnetly carefully crafted input,

trillions of extra checked values, which might be a problem.