Day 20. Fun problem, thew me ffor a slight loop but easy enough.
At first, I obviously thought the maze was going to be more pathfinding, but not so
While I considered encoding the first problem as a graph, were stepping into or out
of a wall would be encoded as going from (x, y, 0) to (x, y+1, 1) to (x, y+2, 2)
and then encoding all the walls as existing on the z=1 plane, I realized that this
would be way overkill, and result in a graph much bigger than what was needed.

_Especially_ since the problem didng't want me to find _a_ path, but many.

The first part is kinda brute-forced by trying to walk in all directions. The second,
using the Manhattan distance and only checking between actual points of the path,
feels very nice in comparison. That could could easily be adapted to solve part 1
cleaner. I'm not gonna, but I could.

My first thought on part 2 was that I was gonna do a sliding window over the maze,
but, gee, that would have been increidbly ineifficient.

Even as it stands, my solution isn't particularly fast. It checks, for any two
points on the path, whether they are within cheating range, then to check if
using the cheat here would save enough time. The checks are blazingly fast, but
it's still being done on approximately 100 Million pairs.

There could probably be a good deal of time saved by only checking path entries that
are >75ish steps ahead of the current one. Same complexity class, but cutting the
runtime in less than half is still meaningful on the ~1 minute runtime.
Would be basically free by just using the path indices. I'm sure there's another even
smarter approach, somehow getting numpy to compute all distances between points in
a parallel, turbo-efficient way?

Well, I got my answers and I'm satisfied.